更新代码

+++ app/src/main/java/com/kekemei/kekemei/bean/BaseBean.java
@@ -42,12 +42,12 @@ public class BaseBean implements Serializable {
     private String project_category_id;
     private String image;
     private String images;
-    private int price_market;
-    private int price_discount;
+    private float price_market;
+    private float price_discount;
     private String treatment_price;
-    private int price_newmember;
-    private int price_member;
-    private int price_second;
+    private float price_newmember;
+    private float price_member;
+    private float price_second;
     private String treatment_count;
     private String quality;
     private String content;
@@ -56,7 +56,7 @@ public class BaseBean implements Serializable {
     private String state;
     private int weigh;
     private int views;
-    private int createtime;
+    private float createtime;
     private double distance;
     private String quality_text;
     private String state_text;
@@ -103,19 +103,19 @@ public class BaseBean implements Serializable {
         this.images = images;
     }
 
-    public int getPrice_market() {
+    public float getPrice_market() {
         return price_market;
     }
 
-    public void setPrice_market(int price_market) {
+    public void setPrice_market(float price_market) {
         this.price_market = price_market;
     }
 
-    public int getPrice_discount() {
+    public float getPrice_discount() {
         return price_discount;
     }
 
-    public void setPrice_discount(int price_discount) {
+    public void setPrice_discount(float price_discount) {
         this.price_discount = price_discount;
     }
 
@@ -127,27 +127,27 @@ public class BaseBean implements Serializable {
         this.treatment_price = treatment_price;
     }
 
-    public int getPrice_newmember() {
+    public float getPrice_newmember() {
         return price_newmember;
     }
 
-    public void setPrice_newmember(int price_newmember) {
+    public void setPrice_newmember(float price_newmember) {
         this.price_newmember = price_newmember;
     }
 
-    public int getPrice_member() {
+    public float getPrice_member() {
         return price_member;
     }
 
-    public void setPrice_member(int price_member) {
+    public void setPrice_member(float price_member) {
         this.price_member = price_member;
     }
 
-    public int getPrice_second() {
+    public float getPrice_second() {
         return price_second;
     }
 
-    public void setPrice_second(int price_second) {
+    public void setPrice_second(float price_second) {
         this.price_second = price_second;
     }
 
@@ -215,11 +215,11 @@ public class BaseBean implements Serializable {
         this.views = views;
     }
 
-    public int getCreatetime() {
+    public float getCreatetime() {
         return createtime;
     }
 
-    public void setCreatetime(int createtime) {
+    public void setCreatetime(float createtime) {
         this.createtime = createtime;
     }
